[Garden] - feat: add last step option to fieldsettings (#1096)
