+++ src/components/ParkView/Components/GroupBySelectors/groupBy.styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const SelectRowWrapper = styled.div`
+    display: flex;
+    align-items: center;
+`;
+
+export const Separator = styled.p`
+    margin-right: 0.5em;
+    margin-left: 0.5em;
+`;
+
+export const SelectOneWrapper = styled.div`
+    width: 250px;
+`;
