+++ src/components/ParkView/Models/fieldSettings.ts
@@ -10,6 +10,8 @@ export type FieldSetting<ItemType> = {
     label?: string;
     getKey?: GetKeyFunction<ItemType>;
     getColumnSort?: GetSortFunction;
+    /** Set to true if no other grouping should be available after this */
+    lastStep?: boolean;
 };
 
 /**
