--- src/components/ParkView/Styles/groupingSelector.ts
@@ -1,15 +1,0 @@
-import styled from 'styled-components';
-
-export const SelectRowWrapper = styled.div`
-    display: flex;
-    align-items: center;
-`;
-
-export const Separator = styled.p`
-    margin-right: 0.5em;
-    margin-left: 0.5em;
-`;
-
-export const SelectOneWrapper = styled.div`
-    width: 250px;
-`;
